ub:ubus: Bugfix control software restart features

Created-by: qiushengming
Commit-by: Junlong Zheng;Anzhe Li;Yiyu Liu;Shilong Wu;Ben Li;Yuhao Xiang
Merged-by: slim6882
Description: ub:ubus: Add ummu_map attribute in sysfs
ub:ubus: Fix X86 build error
dma-mapping: benchmark: add support for UB devices
ub:ubus: call ub_host_probe inside register_ub_manage_subsystem_ops
ub:ubus: Support for UB Memory Decoder general layer implementation
...

See merge request: kunpengcompute/anolis-cloud-kernel!62

+++ drivers/vfio/ubus/vfio_ub_intrs.c
@@ -83,7 +83,7 @@ static int vfio_ub_intr_set_vector_signal(struct vfio_ub_core_device *vdev,
 
 	irq_vec = ub_irq_vector(uent, vector);
 
-	/* if already assigned interrupts, just free it and reassigned it later. */
+	/* if already assigned interrupts, just free them and reassign them later. */
 	if (vdev->ctx[vector].trigger != NULL) {
 		irq_bypass_unregister_producer(&vdev->ctx[vector].producer);
 		free_irq(irq_vec, vdev->ctx[vector].trigger);
@@ -145,7 +145,7 @@ static int vfio_ub_intr_set_block(struct vfio_ub_core_device *vdev, unsigned int
 	if ((start >= vdev->num_ctx) || ((start + count) > vdev->num_ctx))
 		return -EINVAL;
 
-	for (i = 0, j = start; i < count; i++, j++) {
+	for (i = 0, j = (int)start; i < (int)count; i++, j++) {
 		int fd = fds ? fds[i] : -1;
 
 		ret = vfio_ub_intr_set_vector_signal(vdev, j, fd);
@@ -215,7 +215,7 @@ static int vfio_ub_intr_irq_trigger(struct vfio_ub_core_device *vdev,
 	if (flags & VFIO_IRQ_SET_DATA_EVENTFD) {
 		int32_t *fds = (int32_t *)data;
 
-		/* if ub device irq has setted, then update irq eventfd handler */
+		/* if ub device irq has been set, update irq eventfd handler */
 		if (irq_is(vdev, index))
 			return vfio_ub_intr_set_block(vdev, start, count, fds);
